Create public-notes-sheet-script
This current functionality allows theater directors to leave feedback through notes for performers, each of which is given a corresponding string which the performers must input under their name on the sheet. On editing the form, the performers will receive feedback through color-coding as to whether they have input a required word and whether they have read all of their required notes.